feat(api): refactor sorting and filter
